+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/webview/in_app_webview/InAppWebView.kt
@@ -561,6 +561,12 @@ class InAppWebView : InputAwareWebView, InAppWebViewInterface {
       )
     }
 
+    customSettings.webAuthenticationSupport?.let { support ->
+      if (WebViewFeature.isFeatureSupported(WebViewFeature.WEB_AUTHENTICATION)) {
+        WebSettingsCompat.setWebAuthenticationSupport(settings, support)
+      }
+    }
+
     if (
       customSettings.requestedWithHeaderOriginAllowList != null &&
       WebViewFeature.isFeatureSupported(WebViewFeature.REQUESTED_WITH_HEADER_ALLOW_LIST)