+++ flutter_inappwebview_forge_platform_interface/lib/src/types/main.dart
@@ -190,6 +190,7 @@ export 'user_script.dart' show UserScript;
 export 'user_script_injection_time.dart' show UserScriptInjectionTime;
 export 'vertical_scrollbar_position.dart'
     show VerticalScrollbarPosition, AndroidVerticalScrollbarPosition;
+export 'web_authentication_support.dart' show WebAuthenticationSupport;
 export 'web_archive_format.dart' show WebArchiveFormat;
 export 'web_authentication_session_error.dart'
     show WebAuthenticationSessionError;